# backend/ParkingBuilding.CoreApi/openapi.json
{
    "openapi": "3.1.1",
    "info": {
        "title": "ParkingBuilding.CoreApi | v1",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:5000/"
        }
    ],
    "paths": {
...
        "/api/core/db-check": {
            "get": {
                "tags": [
                    "DbCheck"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme. Enter your token in the text input below.",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}